Rename analytics.js to umami.js so that it's not blocked
