+++ src/stores/todos.js
@@ -2,7 +2,7 @@ import { writable, derived } from 'svelte/store';
 import { tweened } from 'svelte/motion';
 import * as uuid from 'uuid';
 import pick from '@lib/pick';
-import { trackEvent } from '@lib/analytics';
+import { trackEvent } from '@lib/umami';
 import { STORAGE_KEY_DATA } from '@lib/constants';
 
 const REMOVE_TIMER_DURATION = 2000;
